+++ engines/ags/module.mk
@@ -24,9 +24,7 @@ MODULE_OBJS = \
 	lib/allegro/math.o \
 	lib/allegro/rotate.o \
 	lib/allegro/surface.o \
-	lib/allegro/surface_simd_neon.o \
-	lib/allegro/surface_simd_sse.o \
-	lib/allegro/surface_simd_none.o \
+	lib/allegro/surface_generic.o \
 	lib/allegro/system.o \
 	lib/allegro/unicode.o \
 	lib/std/std.o \
@@ -377,6 +375,17 @@ MODULE_OBJS += \
 	tests/test_sprintf.o \
 	tests/test_string.o \
 	tests/test_version.o
+endif
+
+ifeq ($(SCUMMVM_NEON),1)
+MODULE_OBJS += \
+	lib/allegro/surface_neon.o
+$(MODULE)/lib/allegro/surface_neon.o: CXXFLAGS += $(NEON_CXXFLAGS)
+endif
+ifeq ($(SCUMMVM_SSE2),1)
+MODULE_OBJS += \
+	lib/allegro/surface_sse2.o
+$(MODULE)/lib/allegro/surface_sse2.o: CXXFLAGS += -msse2
 endif
 
 # This module can be built as a plugin
